version checker
introducing the version checker - a version updater spinoff made to compare the checksum in the file and the real files checksum.
The idea is to provide/upload this tool with the mod after the release to allow players to check if their files are not corrupted. So basically ~73 KB telling if you need to download the 2GB+ again